Closes #5560: Refactor AddonsAdapter and CustomViewHolder into feature addons
- Refactors AddonsAdapter to AddonsManagerAdapter in feature addons.
