+++ components/feature/addons/src/main/res/layout/mozac_feature_addons_item.xml
@@ -1,4 +1,5 @@
-<?xml version="1.0" encoding="utf-8"?><!-- This Source Code Form is subject to the terms of the Mozilla Public
+<?xml version="1.0" encoding="utf-8"?>
+<!-- This Source Code Form is subject to the terms of the Mozilla Public
    - License, v. 2.0. If a copy of the MPL was not distributed with this
    - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
 
@@ -91,6 +92,6 @@
             android:layout_alignParentEnd="true"
             android:layout_centerVertical="true"
             android:contentDescription="@string/mozac_feature_addons_install_addon_content_description"
-            android:tint="@android:color/black"
+            android:tint="?android:attr/textColorPrimary"
             android:src="@drawable/mozac_ic_new" />
 </RelativeLayout>
