+++ samples/browser/src/main/java/org/mozilla/samples/browser/addons/AddonsFragment.kt
@@ -9,33 +9,26 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.RatingBar
-import android.widget.TextView
 import android.widget.Toast
-import androidx.annotation.StringRes
-import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.synthetic.main.fragment_add_ons.*
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.launch
 import mozilla.components.feature.addons.Addon
 import mozilla.components.feature.addons.AddonManagerException
+import mozilla.components.feature.addons.ui.AddonsManagerAdapter
+import mozilla.components.feature.addons.ui.AddonsManagerAdapterDelegate
 import mozilla.components.feature.addons.ui.PermissionsDialogFragment
 import org.mozilla.samples.browser.R
-import org.mozilla.samples.browser.addons.AddonsFragment.CustomViewHolder.AddonViewHolder
-import org.mozilla.samples.browser.addons.AddonsFragment.CustomViewHolder.SectionViewHolder
-import org.mozilla.samples.browser.addons.AddonsFragment.CustomViewHolder.UnsupportedSectionViewHolder
 import org.mozilla.samples.browser.ext.components
 
 /**
  * Fragment use for managing add-ons.
  */
-class AddonsFragment : Fragment(), View.OnClickListener {
+class AddonsFragment : Fragment(), AddonsManagerAdapterDelegate {
     private lateinit var recyclerView: RecyclerView
     private val scope = CoroutineScope(Dispatchers.IO)
 
@@ -70,12 +63,15 @@ class AddonsFragment : Fragment(), View.OnClickListener {
         recyclerView.layoutManager = LinearLayoutManager(requireContext())
         scope.launch {
             try {
-                val addons = requireContext().components.addonManager.getAddons()
+                val context = requireContext()
+                val addonCollectionProvider = context.components.addonCollectionProvider
+                val addons = context.components.addonManager.getAddons()
 
                 scope.launch(Dispatchers.Main) {
-                    val adapter = AddonsAdapter(
-                        this@AddonsFragment,
-                        addons
+                    val adapter = AddonsManagerAdapter(
+                        addonCollectionProvider = addonCollectionProvider,
+                        addonsManagerDelegate = this@AddonsFragment,
+                        addons = addons
                     )
                     recyclerView.adapter = adapter
                 }
@@ -87,245 +83,28 @@ class AddonsFragment : Fragment(), View.OnClickListener {
         }
     }
 
-    /**
-     * An adapter for displaying add-on items.
-     */
-    @Suppress("TooManyFunctions", "LargeClass")
-    inner class AddonsAdapter(
-        private val clickListener: View.OnClickListener,
-        addons: List<Addon>
-    ) : RecyclerView.Adapter<CustomViewHolder>() {
-        private val items: List<Any>
+    override fun onAddonItemClicked(addon: Addon) {
+        val context = requireContext()
 
-        private val unsupportedAddons = ArrayList<Addon>()
-
-        init {
-            items = createListWithSections(addons)
-        }
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CustomViewHolder {
-            return when (viewType) {
-                VIEW_HOLDER_TYPE_ADDON -> createAddonViewHolder(parent)
-                VIEW_HOLDER_TYPE_SECTION -> createSectionViewHolder(parent)
-                VIEW_HOLDER_TYPE_NOT_YET_SUPPORTED_SECTION -> createUnsupportedSectionViewHolder(parent)
-                else -> throw IllegalArgumentException("Unrecognized viewType")
-            }
-        }
-
-        private fun createSectionViewHolder(parent: ViewGroup): CustomViewHolder {
-            val context = parent.context
-            val inflater = LayoutInflater.from(context)
-            val view = inflater.inflate(R.layout.addons_section_item, parent, false)
-            val titleView = view.findViewById<TextView>(R.id.title)
-
-            return SectionViewHolder(view, titleView)
-        }
-
-        private fun createUnsupportedSectionViewHolder(parent: ViewGroup): CustomViewHolder {
-            val context = parent.context
-            val inflater = LayoutInflater.from(context)
-            val view = inflater.inflate(R.layout.addons_section_unsupported_section_item, parent, false)
-            val titleView = view.findViewById<TextView>(R.id.title)
-
-            return UnsupportedSectionViewHolder(view, titleView)
-        }
-
-        private fun createAddonViewHolder(parent: ViewGroup): AddonViewHolder {
-            val context = parent.context
-            val inflater = LayoutInflater.from(context)
-            val view = inflater.inflate(R.layout.add_ons_item, parent, false)
-            val iconView = view.findViewById<ImageView>(R.id.add_on_icon)
-            val titleView = view.findViewById<TextView>(R.id.add_on_name)
-            val summaryView = view.findViewById<TextView>(R.id.add_on_description)
-            val ratingView = view.findViewById<RatingBar>(R.id.rating)
-            val userCountView = view.findViewById<TextView>(R.id.users_count)
-            val addButton = view.findViewById<ImageView>(R.id.add_button)
-            return AddonViewHolder(
-                view,
-                iconView,
-                titleView,
-                summaryView,
-                ratingView,
-                userCountView,
-                addButton
-            )
-        }
-
-        override fun getItemCount() = items.size
-
-        override fun getItemViewType(position: Int): Int {
-            return when (items[position]) {
-                is Addon -> VIEW_HOLDER_TYPE_ADDON
-                is Section -> VIEW_HOLDER_TYPE_SECTION
-                is NotYetSupportedSection -> VIEW_HOLDER_TYPE_NOT_YET_SUPPORTED_SECTION
-                else -> throw IllegalArgumentException("items[position] has unrecognized type")
-            }
-        }
-
-        override fun onBindViewHolder(holder: CustomViewHolder, position: Int) {
-            val item = items[position]
-
-            when (holder) {
-                is SectionViewHolder -> bindSection(holder, item as Section)
-                is AddonViewHolder -> bindAddon(holder, item as Addon)
-                is UnsupportedSectionViewHolder -> bindNotYetSupportedSection(holder, item as NotYetSupportedSection)
-            }
-        }
-
-        private fun bindSection(holder: SectionViewHolder, section: Section) {
-            holder.titleView.setText(section.title)
-        }
-
-        private fun bindNotYetSupportedSection(holder: UnsupportedSectionViewHolder, section: NotYetSupportedSection) {
-            holder.titleView.setText(section.title)
-            holder.itemView.setOnClickListener {
-                val intent = Intent(context, NotYetSupportedAddonActivity::class.java)
-                intent.putExtra("add_ons", unsupportedAddons)
-                context!!.startActivity(intent)
-            }
-        }
-
-        private fun bindAddon(holder: AddonViewHolder, addon: Addon) {
-            val context = holder.itemView.context
-            addon.rating?.let {
-                val userCount = context.getString(R.string.mozac_feature_addons_user_rating_count)
-                val ratingContentDescription =
-                    context.getString(R.string.mozac_feature_addons_rating_content_description)
-                holder.ratingView.contentDescription =
-                    String.format(ratingContentDescription, it.average)
-                holder.ratingView.rating = it.average
-                holder.userCountView.text = String.format(userCount, getFormattedAmount(it.reviews))
-            }
-
-            holder.titleView.text =
-                if (addon.translatableName.isNotEmpty()) {
-                    addon.translatableName.translate()
-                } else {
-                    addon.id
-                }
-
-            if (addon.translatableSummary.isNotEmpty()) {
-                holder.summaryView.text = addon.translatableSummary.translate()
-            } else {
-                holder.summaryView.visibility = View.GONE
-            }
-
-            holder.itemView.tag = addon
-            holder.itemView.setOnClickListener(clickListener)
-            holder.addButton.isVisible = !addon.isInstalled()
-            val listener = if (!addon.isInstalled()) {
-                clickListener
-            } else {
-                null
-            }
-            holder.addButton.setOnClickListener(listener)
-
-            scope.launch {
-                val iconBitmap = context.components.addonCollectionProvider.getAddonIconBitmap(addon)
-
-                iconBitmap?.let {
-                    MainScope().launch {
-                        holder.iconView.setImageBitmap(it)
-                    }
-                }
-            }
-        }
-
-        private fun createListWithSections(addons: List<Addon>): List<Any> {
-            val itemsWithSections = ArrayList<Any>()
-            val installedAddons = ArrayList<Addon>()
-            val recommendedAddons = ArrayList<Addon>()
-
-            addons.forEach { addon ->
-                if (addon.isInstalled()) {
-                    if (!addon.isSupported()) {
-                        unsupportedAddons.add(addon)
-                    } else {
-                        installedAddons.add(addon)
-                    }
-                } else {
-                    recommendedAddons.add(addon)
-                }
-            }
-
-            // Add installed section and addons if available
-            if (installedAddons.isNotEmpty()) {
-                itemsWithSections.add(Section(R.string.mozac_feature_addons_installed_section))
-
-                itemsWithSections.addAll(installedAddons)
-            }
-
-            // Add recommended section and addons if available
-            if (recommendedAddons.isNotEmpty()) {
-                itemsWithSections.add(Section(R.string.mozac_feature_addons_recommended_section))
-
-                itemsWithSections.addAll(recommendedAddons)
-            }
-
-            // Add unsupported section
-            if (unsupportedAddons.isNotEmpty()) {
-                itemsWithSections.add(NotYetSupportedSection(R.string.mozac_feature_addons_unsupported_section))
-            }
-
-            return itemsWithSections
+        if (addon.isInstalled()) {
+            val intent = Intent(context, InstalledAddonDetailsActivity::class.java)
+            intent.putExtra("add_on", addon)
+            context.startActivity(intent)
+        } else {
+            val intent = Intent(context, AddonDetailsActivity::class.java)
+            intent.putExtra("add_on", addon)
+            this.startActivity(intent)
         }
     }
 
-    /**
-     * A base view holder.
-     */
-    sealed class CustomViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-        /**
-         * A view holder for displaying section items.
-         */
-        class SectionViewHolder(
-            view: View,
-            val titleView: TextView
-        ) : CustomViewHolder(view)
-
-        /**
-         * A view holder for displaying Not yet supported section items.
-         */
-        class UnsupportedSectionViewHolder(
-            view: View,
-            val titleView: TextView
-        ) : CustomViewHolder(view)
-
-        /**
-         * A view holder for displaying add-on items.
-         */
-        class AddonViewHolder(
-            view: View,
-            val iconView: ImageView,
-            val titleView: TextView,
-            val summaryView: TextView,
-            val ratingView: RatingBar,
-            val userCountView: TextView,
-            val addButton: ImageView
-        ) : CustomViewHolder(view)
+    override fun onInstallAddonButtonClicked(addon: Addon) {
+        showPermissionDialog(addon)
     }
 
-    override fun onClick(view: View) {
-        val context = view.context
-        when (view.id) {
-            R.id.add_button -> {
-                val addon = (((view.parent) as View).tag as Addon)
-                showPermissionDialog(addon)
-            }
-            R.id.add_on_item -> {
-                val addon = view.tag as Addon
-                if (addon.isInstalled()) {
-                    val intent = Intent(context, InstalledAddonDetailsActivity::class.java)
-                    intent.putExtra("add_on", addon)
-                    context.startActivity(intent)
-                } else {
-                    val intent = Intent(context, AddonDetailsActivity::class.java)
-                    intent.putExtra("add_on", addon)
-                    this.startActivity(intent)
-                }
-            }
-            else -> {}
-        }
+    override fun onNotYetSupportedSectionClicked(unsupportedAddons: ArrayList<Addon>) {
+        val intent = Intent(context, NotYetSupportedAddonActivity::class.java)
+        intent.putExtra("add_ons", unsupportedAddons)
+        requireContext().startActivity(intent)
     }
 
     private fun isAlreadyADialogCreated(): Boolean {
@@ -378,12 +157,5 @@ class AddonsFragment : Fragment(), View.OnClickListener {
 
     companion object {
         private const val PERMISSIONS_DIALOG_FRAGMENT_TAG = "ADDONS_PERMISSIONS_DIALOG_FRAGMENT"
-        private const val VIEW_HOLDER_TYPE_SECTION = 0
-        private const val VIEW_HOLDER_TYPE_NOT_YET_SUPPORTED_SECTION = 1
-        private const val VIEW_HOLDER_TYPE_ADDON = 2
     }
-
-    private inner class Section(@StringRes val title: Int)
-
-    private inner class NotYetSupportedSection(@StringRes val title: Int)
 }
